refactor registration update
