+++ src/database.ts
@@ -12,10 +12,7 @@ import {
     EventAttendanceValidator,
 } from "./services/events/events-schema";
 import { RoleValidator, RoleSchema } from "./services/auth/auth-schema";
-import {
-    RegistrationSchema,
-    RegistrationValidator,
-} from "./services/registration/registration-schema";
+import { RegistrationSchema } from "./services/registration/registration-schema";
 import {
     SubscriptionSchemaValidator,
     SubscriptionSchema,
@@ -86,11 +83,7 @@ export const Database = {
         SubscriptionSchema,
         SubscriptionSchemaValidator
     ),
-    REGISTRATION: initializeModel(
-        "registration",
-        RegistrationSchema,
-        RegistrationValidator
-    ),
+    REGISTRATION: mongoose.model("registration", RegistrationSchema),
     NOTIFICATIONS: initializeModel(
         "notifications",
         NotificationsSchema,
