+++ src/services/registration/registration-router.ts
@@ -3,7 +3,7 @@ import { StatusCodes } from "http-status-codes";
 import {
     RegistrationFilterValidator,
     RegistrationValidator,
-} from "./registration-schema";
+} from "./registration-validators";
 import { Database } from "../../database";
 import RoleChecker from "../../middleware/role-checker";
 import { Role } from "../auth/auth-models";
@@ -26,7 +26,6 @@ registrationRouter.post("/save", RoleChecker([]), async (req, res, next) => {
 
         const registrationData = RegistrationValidator.parse({
             ...req.body,
-            userId: payload.userId,
         });
 
         await Database.REGISTRATION.findOneAndUpdate(
@@ -57,7 +56,6 @@ registrationRouter.post("/submit", RoleChecker([]), async (req, res, next) => {
 
         const registrationData = RegistrationValidator.parse({
             ...req.body,
-            userId: payload.userId,
         });
 
         await Database.REGISTRATION.findOneAndUpdate(
