+++ src/services/registration/registration-schema.ts
@@ -1,32 +1,7 @@
-import mongoose from "mongoose";
-import { z } from "zod";
-
-// Zod schema for registration
-const RegistrationValidator = z.object({
-    userId: z.coerce.string().regex(/user[0-9]*/),
-    name: z.string(),
-    email: z.string().email(),
-    university: z.string(),
-    graduation: z.string().nullable().optional(),
-    major: z.string().nullable().optional(),
-    dietaryRestrictions: z.string().array(),
-    allergies: z.string().array(),
-    age: z.number().nullable().optional(),
-    gender: z.string().nullable().optional(),
-    race: z.array(z.string()).nullable().optional(),
-    ethnicity: z.array(z.string()).nullable().optional(),
-    firstGen: z.string().nullable().optional(),
-    hearAboutRP: z.array(z.string()).nullable().optional(),
-    portfolio: z.string().nullable().optional(),
-    jobInterest: z.array(z.string()).nullable().optional(),
-    isInterestedMechMania: z.boolean(),
-    isInterestedPuzzleBang: z.boolean(),
-    hasResume: z.boolean().default(false),
-    hasSubmitted: z.boolean().optional(),
-});
+import { Schema } from "mongoose";
 
 // Mongoose schema for registration
-const RegistrationSchema = new mongoose.Schema({
+export const RegistrationSchema = new Schema({
     userId: { type: String, required: true, unique: true },
     name: { type: String, required: true },
     email: { type: String, required: true, unique: true },
@@ -48,19 +23,3 @@ const RegistrationSchema = new mongoose.Schema({
     hasResume: { type: Boolean, default: false },
     hasSubmitted: { type: Boolean, default: false },
 });
-
-// Partial schema for attendee filter
-const PartialRegistrationValidator = RegistrationValidator.partial();
-
-const RegistrationFilterValidator = z.object({
-    filter: PartialRegistrationValidator,
-    projection: z.array(
-        z.record(PartialRegistrationValidator.keyof(), z.number().min(1).max(1))
-    ),
-});
-
-export {
-    RegistrationSchema,
-    RegistrationValidator,
-    RegistrationFilterValidator,
-};
